Refactor: Organize routes into modular routers
